Move shuffle code into seedshuffle.ts
